requirements.txt
awscli==1.31.4
beautifulsoup4==4.12.2
boto3==1.33.4
botocore==1.33.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.4
docutils==0.16
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.8.2
six==1.16.0
soupsieve==2.5
uritemplate==4.1.1
urllib3==1.26.18
